Fixing lake-creator shell script markers and refactoring the pytests. (#1107)
* Fixing lake-creator shell script markers and refactoring the pytests.

* Cleaning code and adding changelog

* Cleanup common_utils
