+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Added sleep to Orbit SDK based OrbitJob CR job status check
 - Updating README.md
 - Fixed the expiring kubeconfig issue for long living k8s jobs
+- Fixing lake-creator pytest shell script and refactoring lake-creator pytest files
 
 
 ### **Removed**
