+++ test/regressions/scripts/pytest-lake-creator-multithreaded.sh
@@ -3,20 +3,21 @@
 set -e
 
 # Required os env variables. Replace with testing Orbit env details
-# export AWS_ORBIT_ENV=iter
-# export AWS_ORBIT_TEAM_SPACE=lake-creator
+# AWS_ORBIT_ENV, AWS_ORBIT_TEAM_SPACE
 
 # Set the .kube/config with respect to runtime environment
-pytest -k testlakecreator_cleaner test_lake_creator.py
+
+pytest -k lakecreator_cleaner test_lake_creator_cleaner.py
+
 
 # Multi threaded
-pytest -k testlakecreator_zip -n auto test_lake_creator.py
+pytest -k testlakecreator_unzip -n auto test_lake_creator.py
 
-pytest -k testlakecreator_unzip_check test_lake_creator.py
+pytest -k testlakecreator_check_data_files test_lake_creator.py
 
 # Multi threaded
-pytest -k testlakecreator_glue -n auto test_lake_creator.py
+pytest -k testlakecreator_create_glue_tables -n auto test_lake_creator.py
 
-pytest -k testlakecreator_checkglue test_lake_creator.py
+pytest -k testlakecreator_check_glue_tables test_lake_creator.py
 
 pytest -k testlakecreator_lf test_lake_creator.py
